requirements.txt
aiogram
asyncio
python-barcode
Pillow